requirements.txt
fastapi
uvicorn[standard]
pydantic
python-telegram-bot
python-dotenv
requests
beautifulsoup4
httpx